feat: separate render-hook
